Improve pauseMenuDraw 95.82->96.00 (drop (s16)(s32) cast on lbl_803DD824 s16 stores, removes spurious extsh)

+++ src/main/dll/pausemenu.c
@@ -400,14 +400,14 @@ void pauseMenuDraw(int* arg1, int* arg2, int* arg3)
                 val = b38 - b34;
                 *(u8*)((u8*)lbl_803DD824 + 8) = val;
                 x = *(s16*)((u8*)box + 0x14) + *(u16*)((u8*)box + 8) - (val >> 1) - 0x140;
-                *(s16*)((u8*)lbl_803DD824 + 2) = (s16)(s32)(lbl_803DBA8C * (f32)(s32)x + lbl_803E1F34);
+                *(s16*)((u8*)lbl_803DD824 + 2) = lbl_803DBA8C * (f32)(s32)x + lbl_803E1F34;
 
                 box = gameTextGetBox(0x80);
                 gameTextFn_8001628c(0x3cc, 0, 0, &b38, &b34, &b30, &b2c);
                 val = b38 - b34;
                 *(u8*)((u8*)lbl_803DD824 + 0x28) = val;
                 x = *(s16*)((u8*)box + 0x14) + ((val >> 1) - 0x140);
-                *(s16*)((u8*)lbl_803DD824 + 0x22) = (s16)(s32)(lbl_803DBA8C * (f32)(s32)x + lbl_803E1F34);
+                *(s16*)((u8*)lbl_803DD824 + 0x22) = lbl_803DBA8C * (f32)(s32)x + lbl_803E1F34;
 
                 if (lbl_803DD7D8 != 0)
                 {
